+++ packages/holz-stream-backend/vite.config.ts
@@ -1,6 +1,9 @@
+import { builtinModules } from 'node:module';
 import { defineConfig } from 'vite';
 
 export default defineConfig(async () => {
+  const builtins = builtinModules.map((name) => `node:${name}`);
+
   return {
     build: {
       lib: {
@@ -9,6 +12,7 @@ export default defineConfig(async () => {
         formats: ['es', 'cjs'],
       },
       rollupOptions: {
+        external: builtins,
         output: {
           exports: 'named',
         },